[stores/history.js] Since browserHistory is deprecated, we replace it with the history module and export the createBrowserHistory from history.js as we will be using it in multiple places in the future. We create a stores folder for future use and place the history.js file inside.
[reduxstagram.js]
Since Router is deprecated, we will be using ConnectedRouter with redux but that requires further setup of store. We will import BrowserRouter for now to make sure our routing is working right.

Since IndexRoute is deprecated [originally used in reduxstagram.js], we will use Route from react-router-dom in Main.js and have the switching done inside Main.js

[components/Main.js]
To have multiple route switching, we need to wrap the routes with a Switch component. Within the Switch component, we match the exact path to '/' to the PhotoGrid component using the exact attribute and anything that matches '/view/:postId' to the Single component.