[reducers/posts.js, reducers/comments.js] We create a reducer for each pieces of state that we have
[reducers/index.js]
Since redux can only take in a single reducer, we use the combineReducers method from redux to create a rootReducer. routerReducer is deprecated. So in order for redux to know the changing of state, we use the connectRouter method (which appears in the createStore method [store/index.js]) to communicate information between the history and the rootReducer