[stores/index.js] update folder path to rootReducer
[reduxstagram.js]
We bring in the Provider component from react-redux and also the store that we created earlier. Now that we have our history and store pieces made, we can now utilize ConnectedRouter instead of BrowserRouter with Redux. BrowserRouter does not accept history. We use the history prop of ConnectedRouter with our history object and wrap it with the Provider component and pass it the store object through the store prop.