[components/Main.js] Since we are not using the React.cloneElement to pass this.props down to the child of Main, we now have to do it differently by using a render function and passing down the props differently. This will affect the router match prop in the child later on making it empty (side effect).
[components/Photo.js]
react-addons-css-transition-group has been deprecated instead we bring in react-transition-group and use the TransitionGroup component to wrap our CSSTransition components. In the latest version of react-transition-group we can specify the element prop of TransitionGroup to be null so that it doesn't generate a useless element wrapper if we don't have a need for it. For the TransitionGroup component, we created an options object and use the spread operator to spread it into the component as props.