[components/PhotoGrid.js] Update propname of index to i
[components/Photo.js]
attach onClick event to the like button to run the increment action that we passed down via App.js. We pass the index to the increment action creator

[reducers/posts.js]
When an action is triggered, all reducers will run. To ensure the right reducer is executed, we use a switch block to check for the matching case in the action.type.

To update the state, we use a pure function that would not mutate the original state but instead make a copy of the existing state and update just what we need to update (the single post) - functional programming