[components/Main.js] Since we are passing down all the parents props into <Single/> it overwrites the match prop and now it loses the march.params router value from the url variable postId. To get that back, we pass the state of the Route that wraps Single via props and give it the state.match
[components/Single.js]
Here we locate the post's index by matching the each post's code (UID) to the router match params variable postId. With the post index we then can locate the single post. We then pass these information to the Photo component

[components/Comments.js]
Created the component to be brought in to the Single component