[components/Single.js] Passed the parent props down so that the postId can be read in Comments component
[components/Comments.js]
To get information out of the form fields we used the ref attribute and to read the data, we use this.refs however, this.refs will not work because this refers to the window when it is handleSubmit is called. So in order to retain the original reference of this, we use arrow function. After we got the information we need for the add_comment action creator, we call it and pass in the postId, author, and comment