[components/Comments.js] When add_comment action creator is called from the handleSubmit function, it causes the action to be dispatched and the comments reducer to be subsequently executed. We also attached the remove_comment action dispatcher to the onClick handler of the X button
[reducers/comments.js]
Inside the comments reducer, we use postId as a condition to makesure that the comments reducer acts on the correct dispatched action. We further use the reducer composition method to handle the update of the states by shelling off the update of a single comment to another function called postComments. It is only inside of postComments that we check for the action type using the switch statement since we can either post new comments or remove existing ones.